Moved the Toggle class into the web-components folder.
